Move tarball_validation to utils folder
